requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.2.0
psycopg2-binary==2.9.11
dj-database-url==2.2.0
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2026.1
gunicorn==23.0.0
whitenoise==6.8.2
